refactor: separate query parameter logic into standalone buildSearchParams function
